Bunch of surface-level terraining.
Also, change back to doing the background "AO" effect in a more sane
way.  It's really maybe more of a pain than it's worth, unfortuantely.
Godot engine seems to render things front to back which makes sense, but
it makes convenient subtractive blending difficult.  It has subtractive
blending as an option, but it seems to be an odd exception that behaves
differently than one would really like.