Start on next scene, a park.
Also some fixes and refinements to the beach scene.  It's far from
complete, but the plan will be to flesh things out later.

Also, made the cheat/debug input more sane, it uses the actual unicode
char now instead of the key code, so keys which aren't expressable as a
printable character are ignored and don't throw warnings.