Fixed another BG object oddity.
The previous fix was on the right track but didn't account for different
scales.  I'm not 100% sure why this matters, but taking the scale with
the difference between top and bottom fixes it.  This code has gotten
way too complex for me to keep track of all the moving parts, but
hopefully it's all working?

Also tweak the palette of the above ground concrete/tarmac colors to
mesh a bit better, while still differentiating.