Updated Godot to 4.0.3.
Accounts for all the changed files, which just seem to be adding an
extra parameter.

Also did a bunch of fleshing out of the start area.  It's probably
basically "done".  Maybe as more and more assets get added to the game,
things will change and be swapped in and out, but for now it looks
reasonably "occupied".  Also added all the brick texture variations, so
the walls look less bland.