Establish fairly consistent door/storey scale.
Mostly chosen to be somewhat consistent with the rest of the drawn
graphics so as to avoid having to draw more graphics.  Previously there
wasn't a ton of thought to scale but the gas station followed by the
building outside required for some consistency to be established to
allow a decent level of detail.  Objects in the store are probably a bit
exaggerated scale but that's OK.  Going more for clarity than realism.